+++ front/src/components/Navbar.js
@@ -2,7 +2,7 @@ import Container from "react-bootstrap/Container"
 import Navbar from "react-bootstrap/Navbar"
 import Nav from "react-bootstrap/Nav"
 import { Link } from "react-router-dom"
-import "./style/navbar.css"
+import "../style/navbar.css"
 
 
 export default function NavbarC (props) {
@@ -17,7 +17,8 @@ export default function NavbarC (props) {
           <Nav variant="pills" className="nav nav-pills red">
             <Nav.Link as={Link} to="/Sources" className="whiteText">Filters</Nav.Link>
             <Nav.Link as={Link} to="/Contact" className="whiteText">Contact Us</Nav.Link>
-            <Nav.Link as={Link} to="/" className="whiteText" onCanPlay={()=>{props.setLoginPass(false)}}>Log Out</Nav.Link>
+            <Nav.Link as={Link} to="/" className="whiteText" onClick={()=>{props.setLoginPass(false)}}>Log Out</Nav.Link>
+            <Nav.Link as={Link} to="/SingUp" className="whiteText">Sing Up</Nav.Link>
             </Nav>
         </Navbar.Collapse>
                 </Container>
