--- front/src/components/behind.js
@@ -1,9 +1,0 @@
-import React from "react";
-
-export default function Behind(){
-    return(<><div className="Behind">
-        <footer>
-        @Lorem ipsum dolor sit amet, consectetuer adipiscing elit, sed diam nonummy nibh euismod tincidunt ut 
-        laoreet dolore magna aliquam erat volutpat. Ut wisi enim ad minim veniam, quis nostrud 
-       </footer></div></>)
-}
