+++ front/src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
+import App from './components/App';
 import "./style/login.css"
 import "./style/filters.css"
 import "./style/main.css"
