--- front/src/objects.json
@@ -1,38 +1,0 @@
-[
-   {
-      "Username": "booker12",
-      " Identifier": "9012",
-      "One-time password": "12se74",
-      "Recovery code": "rb9012",
-      "new-obj":
-         {
-           "date": "2020-01-05",
-           "customerId": "11091700",
-           "amount": "3"
-         }
-   },
-   {
-      "Username": "grey07",
-      " Identifier": "2070",
-      "One-time password": "04ap67",
-      "Recovery code": "lg2070",
-      "new-obj":
-         {
-           "date": "2020-01-05",
-           "customerId": "11091700",
-           "amount": "3"
-         }
-   },
-   {
-      "Username": "johnson81",
-      " Identifier": "4081",
-      "One-time password": "30no86",
-      "Recovery code": "cj4081",
-      "new-obj":
-         {
-           "date": "2020-01-05",
-           "customerId": "11091700",
-           "amount": "3"
-         }
-   }
-]
