--- front/src/style/body.css
@@ -1,133 +1,0 @@
-
-.tableView {
-    position: relative;
-    left: 6vw;
-    margin-top: 10px;
-}
-
-.header1 {
-display: flex;
-flex-direction: row;
-width: 1400px;
-height: 120px;
-justify-content: flex-end;
-border-bottom: solid 1px rgb(50, 50, 50);
-}
-
-.linkFirst {
-    color: white;
-}
-
-.header1 img{
-    width:492px;
-    height: 87px;
-}
-
-.backCover {
-    width: 1641px;
-    height: 850px;
-    background: rgb(63,251,72);
-    background: radial-gradient(circle, rgba(63,251,72,1) 92%, rgba(72,83,72,1) 100%);
-    position: absolute;
-    z-index: 900;
-    border-radius: 30px;
-    box-shadow: 15px 15px 20px 3px rgba(0, 0, 0, 0.7);
-}
-
-.frontCover {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: space-evenly;
-    width: 1641px;
-    height:850px;
-    background-color: black;
-    border-radius: 30px 50px 30px 50px;
-    color:white;
-    position: absolute;
-    z-index: 950;
-}
-
-.filterInput {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    width: 1200px;
-}
-
-
-.justInputs{
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    justify-content: space-between;
-    height: 400px;
-}
-
-.labels {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    justify-content: space-between;
-    height: 380px;
-    margin-right: 10px;
-}
-
-.labels label {
-    font-size: 20px;
-    font-weight: 900;
-}
-
-.justInputs input {
-border: black;
-font-size: 24px;
-text-align: center;
-width: 360px;
-height: 40px;
-border-radius: 20px;
-box-shadow: inset 0px 0px 4px 4px black;
-margin-right: 290px;
-}
-
-.buttonSec {
-    display: flex;
-    flex-direction: column;
-}
-
-.buttonSec button {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    width: 300px;
-    height: 65px;
-    border-radius: 20px;
-    box-shadow: inset 0px 0px 4px 4px black;
-    font-size: 30px;
-    font-family: "Suisse Int'l";
-    text-shadow: 1px 1px 2px black;
-    color: white;
-    background: #39C463;
-    margin-bottom: 40px;
-    font-weight: 600;
-}
-
-.buttonSec button[name="secondB"]{
-    background: #39C463;
-    font-weight: 600;
-}
-
-
-.table2 button[name="ok"]{
-    width: 150px;
-    height: 35px;
-    border-radius: 10px;
-    box-shadow: inset 0px 0px 4px 4px black;
-    font-size: 12px;
-    font-family: "Suisse Int'l";
-    text-shadow: 1px 1px 2px black;
-    color: white;
-    background: #39C463;
-    font-weight: 600;
-}
