+++ front/src/style/contact.css
@@ -25,6 +25,7 @@
     transition: 1s !important;
 }
 
+
 .singleCard:hover {
     background-color: rgb(43, 43, 43) !important;
     top: -10px;
