+++ front/src/style/footer.css
@@ -1,5 +1,7 @@
 .footerDiv {
+
     width: 100%;
     text-align: center;
-    margin: 1rem 0;
+    margin-top: 4rem;
+
     }
